+++ examples/info.rs
@@ -20,21 +20,21 @@ fn main() {
 
 	let platforms = Platform::list();
 
-	let context = Context::new(None, None).unwrap();
-	let queue = Queue::new(&context, None);
-	let buffer = Buffer::<f32>::new(&dims, &queue);
-	// let image = Image::new();
-	// let sampler = Sampler::new();
-	let program = Program::builder().src(SRC).build(&context).unwrap();
-	let device = program.device_ids_raw()[0];
-	let kernel = Kernel::new("multiply", &program, &queue, dims.work_dims()).unwrap()
-        .arg_buf(&buffer)
-        .arg_scl(10.0f32);
-    let mut event_list = EventList::new();
+	// let context = Context::new(None, None).unwrap();
+	// let queue = Queue::new(&context, None);
+	// let buffer = Buffer::<f32>::new(&dims, &queue);
+	// // let image = Image::new();
+	// // let sampler = Sampler::new();
+	// let program = Program::builder().src(SRC).build(&context).unwrap();
+	// let device = program.device_ids_raw()[0];
+	// let kernel = Kernel::new("multiply", &program, &queue, dims.work_dims()).unwrap()
+ //        .arg_buf(&buffer)
+ //        .arg_scl(10.0f32);
+ //    let mut event_list = EventList::new();
 
-    kernel.enqueue(None, Some(&mut event_list));
-    let event = event_list.last().unwrap().clone();
-    event_list.wait();
+ //    kernel.enqueue(None, Some(&mut event_list));
+ //    let event = event_list.last().unwrap().clone();
+ //    event_list.wait();
 
     for platform in platforms {
     	println!("{}", platform);