+++ src/raw/custom/mod.rs
@@ -1,1 +1,2 @@
 pub mod enums;
+pub mod structs;